feat: config database provisioning via supabase
