+++ .gitignore
@@ -1,9 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
-# ! while still in dev, ignore migrations
-/ignore/drizzle/migrations
-
 # dependencies
+/supabase/.temp
 /node_modules
 /.pnp
 .pnp.*
