+++ app/api/todos/route.ts
@@ -1,20 +1,18 @@
 import { NextRequest, NextResponse } from "next/server";
-import { createClient } from "@/lib/supabase/server";
-import { getTodosByUser, createTodo, updateTodo, deleteTodo } from "@/services/todoService";
-
-// ! THIS NEEDS ZOD VALIDATION
+import { createClient } from "@/supabase/server";
 
 export async function GET() {
   const supabase = await createClient();
   const {
     data: { user },
   } = await supabase.auth.getUser();
-
   if (!user) return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
 
-  const todos = await getTodosByUser(user.id);
+  // RLS automatically filters by owner_id = user.id
+  const { data, error } = await supabase.from("todos").select("*");
+  if (error) return NextResponse.json({ error: error.message }, { status: 400 });
 
-  return NextResponse.json(todos);
+  return NextResponse.json(data);
 }
 
 export async function POST(req: NextRequest) {
@@ -25,8 +23,20 @@ export async function POST(req: NextRequest) {
   if (!user) return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
 
   const body = await req.json();
-  const todo = await createTodo(user.id, body.title, body.content);
-  return NextResponse.json(todo);
+
+  const { data, error } = await supabase
+    .from("todos")
+    .insert({
+      title: body.title,
+      content: body.content,
+      owner_id: user.id, // RLS requires this
+    })
+    .select()
+    .single();
+
+  if (error) return NextResponse.json({ error: error.message }, { status: 400 });
+
+  return NextResponse.json(data);
 }
 
 export async function PUT(req: NextRequest) {
@@ -37,8 +47,22 @@ export async function PUT(req: NextRequest) {
   if (!user) return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
 
   const body = await req.json();
-  const todo = await updateTodo(user.id, body.id, body);
-  return NextResponse.json(todo);
+
+  const { data, error } = await supabase
+    .from("todos")
+    .update({
+      title: body.title,
+      content: body.content,
+      done: body.done,
+    })
+    .eq("id", body.id)
+    // RLS enforces owner_id = user.id automatically
+    .select()
+    .single();
+
+  if (error) return NextResponse.json({ error: error.message }, { status: 400 });
+
+  return NextResponse.json(data);
 }
 
 export async function DELETE(req: NextRequest) {
@@ -49,6 +73,10 @@ export async function DELETE(req: NextRequest) {
   if (!user) return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
 
   const { id } = await req.json();
-  const todo = await deleteTodo(user.id, id);
-  return NextResponse.json(todo);
+
+  const { data, error } = await supabase.from("todos").delete().eq("id", id).select().single();
+
+  if (error) return NextResponse.json({ error: error.message }, { status: 400 });
+
+  return NextResponse.json(data);
 }
