+++ components/auth-button.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { createClient } from "@/lib/supabase/client";
+import { createClient } from "@/supabase/client";
 import { LogoutButton } from "./logout-button";
 import { User } from "@supabase/supabase-js";
