+++ components/password-login.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { createClient } from "@/lib/supabase/client";
+import { createClient } from "@/supabase/client";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { Label } from "@/components/ui/label";
