--- drizzle/index.ts
@@ -1,8 +1,0 @@
-import postgres from "postgres";
-import { drizzle } from "drizzle-orm/postgres-js";
-
-const connectionString = process.env.DATABASE_URL as string;
-
-// Disable prefetch as it is not supported for "Transaction" pool mode
-export const client = postgres(connectionString, { prepare: false });
-export const db = drizzle(client);
